Can not get semantic memory with MCP (#815)
semantic memory does not work for mcp

The semantic memory returned in mcp search memory response
is always empty.

The root cause is that semantic memory does not support metadata filter.
In order for semantic memory to work, just query the information
in the whole project in mcp.  The good part is that in mcp, the project
is used by only one user.  So the solution should work for now.